requirements.txt
attrs==22.1.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.0.3
Flask-WTF==1.0.1
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
platformdirs==2.5.4
pyparsing==3.0.9
python-dotenv==0.21.0
pythonfinder==1.3.1
requests==2.28.1
urllib3==1.26.13
virtenv==0.3.1
virtualenv==20.17.0
Werkzeug==2.2.2
WTForms==3.0.1
